Windows FSocket/FThread updated
git-svn-id: https://client-sdk.forge.funambol.org/svn/client-sdk/branches/cpp-sdk/us1408@401 3966be30-755a-0410-815e-dd79f492ac3c

+++ cpp-sdk/src/include/windows/base/winadapter.h
@@ -63,7 +63,7 @@ typedef __int32            int32_t;
 typedef __int64            int64_t;
 typedef unsigned __int8    uint8_t;
 typedef unsigned __int16   uint16_t;
-typedef unsigned __int32   uint16_t;
+typedef unsigned __int32   uint32_t;
 typedef unsigned __int64   uint64_t;
 
 // Define this to the implementation of the sprintf variant with
